Move sc_setup_mount_profiles call to after critical confinement.
This patch moves all of the mount profile processing to happen after we
pivot_root (but only on classic, this is not happening on non-classic)
and always after unshare(CLONE_NEWNS).

This way mount profiles can do anything they way but whatever mount
operations are applied, none of them break out to the main system.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>