use execle() with empty environment when calling snappy-app-dev
use execle() with clean environment instead of execl() when calling
snappy-add-dev to avoid influencing snappy-add-dev and allowing privilege
escalation. snappy-app-dev is only called when devices are assigned via udev
tagging, but currently snapd does not create these udev tags during 'snap
install' or 'snap connect' so this code branch is not entered during app
invocation.

Currently this is not a security issue since no snaps will be udev tagged via
snapd (as such, today this can only be exploited by manually creating the
necessary udev tags (and there is no reason or documentation to do so), but
this requires root and is thus not a privilege escalation).

This is a vulnerability in waiting though since whenever the udev tagging is
used and generated by snapd, then an unprivileged user would be able to create
a crafted environment and pass to snappy-app-dev via execl(). While
snappy-app-dev is constrained by the launcher AppArmor profile, clearly this
issue must be addressed.

Thanks to Sebastian Krahmer from SUSE for discovering this issue.