snap/squashfs: fix test not to hardcode snap size
It turns out that mksquashfs on some distributions is patched to have a
different block size to what we assumed when we wrote the test on
Ubuntu. Instead of being very specific about it, let's just compare the
size of hashed data with the size of the file on disk.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>