Add configure option to select merged /usr FHS
On some distributions the /usr directory is now "merged" into the root
filesystem. In that layout symlinks provide the typical FHS view but
some things are moved around to incompatible locations.

One of the directories affected by this is /media. In the merged FHS it
is now buried somewhere in /run. In the prior layout it is in, well,
/media itself.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>