Bind mount /var/log for log-observe
This patch bind mounts the /var/log directory so that snaps can use the
log-observe interface to get access to system log files.

The fix also involves another permission in the apparmor profile for the
particular mount to go ahead.

The whole set is tested with a new spread regression test.

Fixes: https://bugs.launchpad.net/snap-confine/+bug/1606277
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>