update code for the snap_mode
We updated the bootloader logic in 16.04 which in incompatible to
what we had in 15.04. The new logic is:
 snap_mode={"","try","trying"}
 snap_{os,kernel}="snap-name.snap"
 snap_try_{os,kernel}="snap-try-name.snap"

The bootloaders gadget snaps need to get updated so that they look
at the right boot variable. When a new kernel/core is installed,
the snap_mode is set to "try". The bootloader will then set the
mode to "trying". The snapd.booted code clears the snap_mode
environment. If the bootloader encounters a "snap_mode=trying"
it knows the boot failed and it will boot the previous good kernel.