Implement a fuse interface
I wrote that while working on the LXD snap. We won't actually be needing
it but figured it'd be useful to others.

On a normal system everyone can access /dev/fuse and use it, it's
therefore assumed that this is a perfectly safe kernel feature for
unprivileged users to use so I don't see any reason to restrict this
particular interface.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>