add dbus-app interface (LP: #1590679)
This implements dbus-app in support of 'Apps can't own session bus names'. An
exploratory PR is found in: https://github.com/snapcore/snapd/pull/1446 but
that PR was more than what is needed and it was decided we should implement a
subset of that to address the bug, and then revisit once more details are
available.

In this PR, we allow dbus binding to a well-known name via 'slots':

  name: gnome-logs
  slots:
    interface: dbus-app
    session:
    - org.gnome.Logs

We use 'slots' instead of plugs because the application performing the bind on
dbus is providing a service end point for other dbus processes to communicate
with. In this implementation, on all-snaps systems we do not allow any
communications to the slot implementation (ie, it may only bind, not send or
receive) and on classic systems we allow communications from unconfined
processes to the slot implementation. In this manner, the snap using this
interface can bind to the dbus well-known name without issue (permanent slot
is unconditionally granted) and it may operate as a leaf application in
all-snaps and classic environments.

Since we know that some applications bind to multiple well-known names, an app
is allowed to bind to any number of names.

Since well-known DBus names are rather free-form and don't typically conform to
the snap name and because snaps sometimes need to bind to multiple well-known
names, the store will ensure that snaps from different publishers do not
collide. Future assertions work will allow for auto-approval in the store once
the assertion is granted for binding to the well-known name. In the short term,
we will adjust the review tools to do an easy check (if snapname is 'foo' and
bind interface is 'bar', auto-approve, else manual).

t#