unconditionally mount /usr/src (LP: #1597842)
Some snaps need access to the kernel sources. On an Ubuntu classic system these
are found in /usr/src. This does not cover other locations for cross-distro
or kernel snaps.

To unblock people, this simply mounts /usr/src with the other OS mounts, but
this should really be interface-dependent along with /var/log. Since that
requires a bit of reorganizing, https://bugs.launchpad.net/snappy/+bug/1609499
covers this work.