Do not install a snap if it is already valid
On firstboot we may call snap.Install() for the core/kernel snap
that is in the seed/snaps directory so that it gets copied to
the snaps directory. We do not need to copy in this case because
the files are already identical. The copy is actually harmful
because the system will continue to read from the squashfs
ubuntu-core file while it is being copied which leads to errors
during the boot process.