Use downstream packaging in spread tests
This patch changes spread setup phase to build the native package (for
Ubuntu and soon for Debian) in a way that is more consistent with the
actual release process. This should help us release with confidence as
all CI tests will run with real downstream packaging applied to real
upstream tarball and build in an isolated manner with sbuild.

This also paves the way towards the removal of the debian/ packaging
from the upstream repository. This will be done in a separate pull
request to avoid clutter.

The spread prepare stage was moved to .spread-prepare.sh so that it can
be shell-checked more easily.

The debian-8 spread target was removed because currently the downstream
packaging is not compatible. Debian maintainers were notified to the
issue so that we should be able to re-enable Debian (sid) support soon.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>