Drop ubuntu-core-snapd-units package, use release.OnClassic instead
Merge the "dangerous" systemd services back into the main snapd
package. Instead of an extra package the code will simpliy check
for release.OnClassic() for dangerous bits like "booted" or when
enabling the first ethernet.

This allows us to run the same firstboot seed.yaml population on
a classic system on first-boot (or first-install of snapd) which
allows us to pre-install ubuntu-core to avoid the extra download
on the first snap install.