many: split public snapd REST API into separate socket.
Currently snapd uses a single socket for REST API communication. This
means that, even if a snap wants to access a harmless piece of the API,
it needs to request the snapd-control interface, which is reserved and
opens up the entire API, even the sensitive parts.

This commit is an attempt to split the public (i.e. non-sensitive) bits
of the REST API into a separate socket, access to which is allowed in
the default profiles. Currently the only public API is that which is
used for snapctl.

Updates: #1586465

Signed-off-by: Kyle Fazzari <kyle@canonical.com>