Support checking account-key-request assertions
account-key-request assertions lack an authority-id and are self-signed,
but we want to be able to add them to the database anyway as a record of
the request that caused us to create an account-key.  Add a separate set
of checkers for no-authority assertions and use them in this case.

Although there are some commonalities between these and normal checkers,
they are intentionally separate in order that developers working on
normal assertions don't have to take such careful account of how changes
affect the unusual self-signed cases.