[DONT MERGE] Pass slot/plug into AutoConnect and move special logic.
Instead of putting extra logic in AutoConnectCandidates to determine if we
should automatically connect things like the content interface, etc. Put
this logic into the AutoConnect function itsself. In order to support this
change the parameters this interface function takes and adjust all relevant
built-in interfaces and unit tests.

FIXME: for now the content tests inside interfaces/repo_test.go are removed.
These need to be reintegrated back into interfaces/builtin/content_test.go.
The challenge will be if we can setup a 'repo' to test in.