Combine mount namespace setup into one function
This patch cleans up the API exposed by mount-support.h to include just
the creation and population of the mount namespace. This is done in anticipation
of the upcoming mount namespace sharing feature.

Various things were moved around but there should be no observable difference
between what is done now and what was done before.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>