snap: lessen annoyance of implicit interface tests
The existing interface tests check for the precise count of implicit
slots defined on the core snap. This has to be adjusted with every
single added interface. To lessen the number of files that have to be
edited, let's just check that there are some implicit slots and let test
developers decide if they want to add an explicit check for a particular
interface name.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>