Add the snap-discard-ns executable
This executable uses the namespace group API to discard a possibly
existing mount namespace. The program is not run as setuid root and thus
has no apparmor profile at this time.

The dependencies could be a little bit weaker. It pulls in libseccomp
just because it is used by cleanup function. It also pulls in
libapparmor because of the hat changing logic in ns-support.c. This can
be improved later as it has little cost.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>