Discard snapd-hacker-toolbelt namespaces on restore
Because tests assume pristine test environment and preserved mount
namespaces are not yet discarded by snapd on the removal of a given snap
(branch pending but not merged) we just do it by hand.

Since not all of the tests currently use the snapd-hacker-toolbelt snap
the cleanup code is allowed to fail by ignoring any errors.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>