reorganize to create docker-support and docker interfaces
Shuffle policy around to docker-support and docker interfaces to clarify
connections.

The docker interface is for client snaps to access snaps providing the docker
interface. Slot implementations are reserved for the Docker project.

The docker-support interface is modelled after lxd-support and is for snaps
providing the docker interface to access resources and syscalls necessary to run Docker application
containers. Providing snaps specifying this interface currently may only be
established with the Docker project.