Rework mount namespace support
This patch changes the contents of the initialized mount namespace. The
main motivator is to allow /media to be shared amongst the initial mount
namespace and the derivative namespaces created by snap-confine.

The code is unified and simplified around the new boostrap function. It
now takes a few simple arguments through a configuration structure. The
main argument is the location of the starting point of the desired root
filesystem. On classic this is the core snap (or ubuntu-core if core is
not available). On an all-snap system this is the root filesystem.

The bootstrap function puts the desired root filesystem into a new
temporary directory (using recursive bind mount) paying special
attention to avoid creating bind mount loops. This directory is now
called the constructed root filesystem and eventually, this is where we
pivot_root into.

Since we still have access to the initial root filesystem under the
initial sharing parameters as well as to the new constructed root
filesystem, we can choose to bind mount arbitrary directories over.

The code now supports two lists of directories, one that is bind-mounted
rslave (called unidirectional because events propagate only in one
direction in the peer group) and one that is mounted rshared (called
bidirectional for the obvious reason).

In this patch /media is on the bidirectional list. On classic a large
list of directories is on the unidirectional list. On all snap the
unidirectional list is empty because we don't need anything as we're
sharing the root filesystem which is already correctly populated by the
all-snap initrd.

The existing code for special treatment of /snap, /etc/alternatives,
/tmp and /dev/pts is retained as is (semantically) but has been folded
into the bootstrap function for simplicity.

The code for nvidia support is now also handled by the bootstrap
function but this should be moved outside later. This can be done only
after nvidia code is adjusted to assume it executes after pivot root
(this will simplify it and allow us to do something sensible in all-snap
system later). I chose not to do this to simplify the change and review
process. To repeat, nvidia support is exactly as it was before.

The apparmor profile was adjusted to take account of all the new
(numerous) operations.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>