interfaces,docs: allow sharing SNAP{,_DATA,_COMMON} via content iface
This patch expands the features of the content interface to allow
sharing files from either $SNAP (the snap content), $SNAP_DATA (the
per-revision snap data) or $SNAP_COMMON (the data common across
revisions).

The old syntax, without any of those variables behaves as if $SNAP was
implicitly used so backwards compatiblity is retained.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>