overlord/interfaces: use new Disconnect/DisconnectAll
This patch changes the interface manager to use the new APIs offered by
interfaces.Repository for disconnecting. In particular we no longer
obtain snap.Info of the affected snaps from the repository (which can
contain stale data). Instead we use the snapstate to fetch the current
info).

In addition of the fully-spelled-out "disconnect snap:plug snap:slot" we
now correctly support "disconnect snap:plug" or "disconnect snap:slot".
In those two new forms the snap name can be empty (it then defaults to
the core snap). This is thanks to the use of the new DisconnectAll
method.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>