cmd/snap: generate account-key revocation requests
Add "snap export-key --account=<account-id> --revoke" to generate an
account-key revocation request, which is formatted as an
account-key-request with since == until.  We consult the store to
determine the necessary revision of the new account-key-request
assertion.

Note that the client may not yet always respond to revoked assertions
correctly, e.g. by removing packages signed by a revoked key, but it
will at least refuse to accept the revoked key for signing future
assertions.