Set PATH unconditionally even if mount namespace is set up
This patch corrects a bug where PATH would be correctly set only on the
first invocation of a given snap's application. After the first
invodation the mount namespace would be already initialized and the
relevant code that sets PATH would not run. Now this is corrected and
PATH is always set.

Fixes: https://bugs.launchpad.net/ubuntu/+source/snap-confine/+bug/1630479
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>