drop 'owner' check on mountinfo and allow write to @{PROC}/[0-9]*/attr/current
Due to a kernel bug, the ouid is not being set correctly for /proc accesses by
setuid processes running in user namespaces. While the kernel needs to be
fixed, drop the 'owner' match on @{PROC}/*/mountinfo for now.

Since we are using aa_change_hat(), snap-confine needs to be able to write to
@{PROC}/[0-9]*/attr/current.

Bug: https://launchpad.net/bugs/1630789