Add spread tests for mount namespace layout
This patch adds a simple test that looks at /proc/self/mountinfo within
confined applications. The test discards some information that is too
variable to be useful but retains the most essential facts.

This test will be changed by the upcoming /media sharing patches. It is
intended to be useful for before/after comparison.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>