spread.yaml: dump apparmor denials on spread failure
We're facing an intermittent failure that manifests itself as a failure
to one of the mount operations in snap-confine. To debug the problem
we'd like to dump apparmor denials each time a spread test fails. This
is useful in travis, where the log is kept around and where we cannot
log in interactively to investigate.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>