interfaces: add Snippet interface type
This patch adds a new interface for typed snippets. The idea is that
instead of returning snippets as strings/bytes that have to be parsed by
the backend we can actually pass typed data that the backend
understands. This patch just adds a go interface called Snippet that
represents the opaque snippet.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>