Use sysd.{Disable,Stop} instead of sysd.DisableNow()
We currently have systemd.{Enable,Disable}Now() functions and also
systemd.{Enable,Disable,Start,Stop}(). To avoid confusion remove
the systemd.{Enable,Disable}Now(). EnableNow is not used at all in
the code. And DisableNow() can be replaced with Disable/Stop. The
benefit is that Stop() actually checks that the service really got
stopped and kills it after the given timeout.