interface: openvswitch-support
Misc updates across a number of existing interfaces to support
the use of openvswitch from within a snap.

Addition on a new interface 'openvswitch-support'; as other interfaces
already provide the required apparmor and seccomp stanza, this interface
is simply used to ensure that the openvswitch kernel module is loaded.

openvswitch can be run in a snap using:

    plugs:
      - network
      - network-bind
      - network-control
      - openvswitch-support
      - process-control
      - system-trace