osutil: make RealUser only look at SUDO_USER when uid==0 (#2283)
The RealUser helper looks for the user behind a sudo invocation. But
we only care about the case of "user->root". Sudo can also be used for
"root->user" in which case SUDO_USER is "root". This leads to errors
if you run (as root): "sudo -u some-user snap login" as it will try
to chown the login data file to the SUDO_USER uid (which is root).

LP: #1638656