Add apparmor support module
This patch adds a new set of files, mimicking the xxx-support pattern
used for other changes, that to some extent abstracts the use of
apparmor in snap-confine.

Even when snap-confine is not compiled with apparmor support the same
APIs are available and gracefully degrade to no-ops.

As a small extension, snap confine can now know if it is confined (e.g.
a development version running in a random directory is not confined) and
will no longer crash when changing hats via aa_change_hat().

The patch doesn't yet change any of the tree to use the new functions.
This will be done in the next commit.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>