overlord/ifacestate: don't use DevModeAllowed
This patch fixes a bug where "devmode" would be applied to any snap that
was actually attempting to use "jailmode". The DevModeAllowed function
was incorrectly used as the API for interface security setup was
somewhat confusing earlier.

The API was changed to much more friendly and effective
ConfinementOptions earlier and now it's the time to use them correctly.
Instead of calling DevModeAllowed the values from SnapState or SnapSetup
(from Flags technically) are simply mapped to their ConfinementOption
equivalents.

Fixes: https://bugs.launchpad.net/snappy/+bug/1641885
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>