Discard snap mount namespace on any content interface related change
The mount namespace will be discarded whenever a snap content plug
is connected/disconnected so that snap-confine generates or removes the
required bind mounts.

Previously, a bind mount would never be generated if a command was invoked from snap
declaring a content plug was installed prior to connecting the content plug.
Similarly, a content i/f related bind mount was not removed after disconnecting the
content plug.