Add error support code
This patch adds die()-like APIs that can carry error information around
the control flow. They are meant to provide richer error information  in
cases where the error is not fatal and needs to be returned to the
caller. The APIs are 100% tested and also integrate with die() where
appropriate().

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>