debian: add split ubuntu-core-launcher and snap-confine packages
This patch changes debian packaging for snapd to generate two additional
binary packages, snap-confine and ubuntu-core-launcher. They are now
generated separately so that apt logic can give everyone a better
upgrade experience as snap-confine and ubuntu-core-launcher packages
will no longer have to be removed during the upgrade operation.

To ensure that everything is in sync there's now a strong dependency on
the exact source package version among snapd and snap-confine and also
among snap-confine and ubunut-core-launcher. This should ensure that
those packages are always updated in lockstep, to the extent possible in
dpkg.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>