Apply security comments to AppArmor rules
The security audit over the first proposal revealed that the AppArmor
rules were not entirely correct as they allowed rw access to
'/run/user/[0-9]*/' which should not be allowed.

This commit changes it in a way that a snap that 'plugs: [pulseaudio]'
has only a read access to this location while still keeping the rw
access to the 'pulse' subdirectory there.