interfaces/apparmor: ignore snippets in classic confinement
Apparmor snippets were designed for strict confinement and contain
rules that may conflict with the base template for snaps under classic
confinement. Because we want to allow people to use interfaces with
their classic-confined snaps to open the door toward strict confinement
we don't want to ban interfaces outright. This patch makes apparmor
ignore all snippets for snaps using classic confinement unless jailmode
is enabled (in jailmode, those snippets are meaningful as the base
template changes to the strict template).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>