cmd/snap-confine: honor SNAP_CONFINE_TESTING=bind-packaged-over-core
This patch allows snap-confine to bind mount packaged versions of
snapctl and snap-exec over whatever is shipping in the core snap (or the
ubuntu-core snap). This effectively allows snap-confine to ensure that
the tested package takes priority over whatever is present in the core
snap without having to re-package the core snap each time.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>