use the debian/*.install files instead of debian/rules
The debian/*.install files are the canonical way to install files
in debian packags. We need to use debian/rules for some bits (like
the systemd stuff that is different on xenial and trusty. However
we should minimize this as much as possible to make debian/rules
easier to read.