interfaces: add new-style interfaces.
This patch begins the transition of the all backends and interfaces to
the new interface API, dubbed NewInterface.

Instead of offering snippets, interfaces can now interact with a
backend-specific recorders to record desired changes to the security
system. Ulinke in the past, there are no opaque strings with internal
structure passed around. Everything is recorded in a fully typed manner.

This has the advantage of being easier to process inside the backend as
there's no need to parse snippets. It also has the potential to be
easier to work with in actual interfaces as the recorders can offer
intuitive methods rather than forcing developers to come up with
syntax-perfect snippet strings.

To prove that this all works correctly the mount security backend and
hte content interface are changed to this new style.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>