cmd/snap: add shell completion to connect
This patch adds somewhat imperfect support for shell completion to "snap
connect". The part that is not so nice yet is that colon seems to
terminate the completion input (in bash at least) unless the user
manually adds quotes around the argument and goes back (after the
initial suggestion) to place the cursor immediately before the closing
double quote but after the colon.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>