interfaces: start the transition of mount backend to new API
This patch begins the transition of the mount security backend and the
single interface using it, content, to the new interface API.

Instead of offering snippets, interfaces can now implement a
backend-specific interface, here called MountAware, to enter into rich
interaction with a specific security backend. Each of the new methods
gets a backend controller object (for lack of a better name) that has
APIs for defining desired security changes. The mount backend uses just
one method, AddMountEntry.

Ulinke in the past, there are no opaque strings with internal structure
passes around. Everything is recorded in a fully typed manner. The
backend can use the collected data to perform its job.

This has the advantage of being easier to process inside the backend as
there's no need to parse snippets. It also has the potential to be
easier to work with in actual interfaces as the backend controller can
offer intuitive methods rather than forcing developers to come up with
syntax-perfect snippet strings.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>