interfaces: add new interface API
This patch begins the transition of the all backends and interfaces to the new
interface API.

Instead of offering snippets, interfaces can now interact with a
backend-specific recorders, using backend-specific methods, to record desired
changes to the security system.

This has the advantage of being easier to process inside the backend as there's
no need to parse snippets. It also has the potential to be easier to work with
in actual interfaces as the recorders can offer intuitive methods rather than
forcing developers to come up with syntax-perfect snippet strings.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>