interfaces: allow reading files from previously installed versions by default
While daemon commands are restarted on upgrade, long running non-daemon
commands are not so when snap refresh is performed and the policy is
regenerated, these long running commands may try to access previous revisions
which has caused issues like LP: #1616650 and some of the denials found in
LP: #1655992. This commit does not fully address the problem with policy
reloads where the revision changes, but does ease the pain with regard to the
INSTALL_DIR. There is no harm in allowing read from previous install
directories, so just allow it by default.