many: detect potentially insecure use of snap-confine
The core snap is the only snap that is allowed to ship setuid-root
executables. Amnong those executables is snap-confine. Typically it is
itself confined with a tailored apparmor profile. Due to the way
apparmor profiles are applied, this is only happening when it is
installed as a system package or when it executes in the core snap on an
all-snap system.

On any system supplied with apparmor that has access to a mounted core
snap, unprivileged users may attempt to attack snap-confine with
malicious input (environment, arguments, etc). To defend against such
attacks snap-confine will now refuse to run if it should be confined and
has elevated permissions but the profile is not applied.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>