overlord/ifacestate: setup security of snaps affected by auto-connection
This patch fixes a bug where snaps that were auto-connected did not
have their security setup done. The only reason that this went unnoticed
for so long is that the main snap (the one that triggered the operation)
was handled separately and that most slots don't need a security setup
for an active connection.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>