interfaces: bultin: add core-support
The core-support interface is especially build to allow certain things
to the core snap only. As the core snap is taken as any other snap for
applications and hooks everything runs strictly confined and with that
we need to have proper interfaces in place for everything a hook or
applications wnats to do.

This new interface will cover the basic bits of what is needed. So far
we grant access to the systemd service to allow service start/stop/
restart and modifications inside /etc/ssh to allow disabling the
system sshd service.