cmd: add non-stub implementation of snap-update-ns
This patch adds an almost full implementation of snap-update-ns.

This new executable includes a simple utility module for working with
mount entries (e.g. /etc/fstab) that is based on the C library for the
hard part (parsing).

The missing bits are:
 - there's no locking; I will tweak ns-support to expose that better
 - snapd doesn't generate the snap.$SNAP_NAME.fstab file yet
 - snapd doesn't call snap-update-ns yet (easy)
 - snap-update-ns cannot yet realize that a mount already happened

The last missing bit requires some more explanation. The tool already
inspect existing mount points but due to the way bind mounts are exposed
in the kernel it is non-trivial to realize that a given place is a bind
mount (even if we have all the information in front of us).

One notable unfortunate fact is that snap-update-ns cannot take advantage
of valgrind for debugging:

--41370-- WARNING: unhandled amd64-linux syscall: 308
--41370-- You may be able to write your own handler.
--41370-- Read the file README_MISSING_SYSCALL_OR_IOCTL.
--41370-- Nevertheless we consider this a bug.  Please report
--41370-- it at http://valgrind.org/support/bug_reports.html.
cannot re-associate with mount namespace of snap unix-domain-client: Function not implemented

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>