cmd/snap-confine: support parsing user names and groups in seccomp policy
Add new syntax when parsing seccomp policy:
 u:USER
 g: GROUP

This will find the uid/gid and use SCMP_CMP_EQ when argument filtering. Eg:

chown - u:root g:shadow

This would do a local lookup and allow 'chown root:shadow ...'. This
functionality should only be used with on-the-system users and groups (as
opposed to LDAP) to avoid slow startup times.

Since it is possible for the user or group to no longer exist on the system,
print a warning and skip the rule when not found.