cmd/libsnap: rewrite buffer management for simplicity
This patch changes the way sc_mount_cmd() formats the remaining mount
option for the "-o ..." part of the command.

It used to use the existing buffer, carefully crafting the new message
at the end of the used space. This was correct but non-obvious without
checking how those functions currently operate. The new code simply
creates the "-o ..." string in a a separate buffer and appends that
extra buffer to the main buffer with sc_string_append.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>