overlord/ifacestate: setup seccomp security on startup
This patch attempts to unbreak the world by automatically refreshing the
seccomp profiles on startp of snapd. Seccomp profiles are not expensive
to compute and don't require compilation or loading.

This lets us fix the problem that surfaced when snapd started generating
a syntax, in the seccomp profile, that only new snap-confine
understands. Because snap-confine does not conceptually re-execute yet
everyone that is on the edge channel is now experiencing problems.

This patch should be applied along the next patch that generates
compatible seccomp profiles, without the new syntax.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>