osutil/build_id: add package for reading Build-ID
This patch adds a small package that reads the GNU Build-ID note from an
elf executable. Build-ID is a standard feature in many distributions,
eixsting ever since Fedora 8 introduced it. It facilitiates crash
analysis as each crash contains identified unique for each build that
can be correlated to debugging symbols.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>