Assume higher version of supported distros will still work
The current code in release.go:os.ForceDevMode() uses a strict
version check to test if a new release supports non-devmode.

This is problematic as we always need to add new versions when
a derivate releases a new version. This branch uses VersionCompare()
to check the version and assumes that higher versions will continue
to work. It also assume somewhat sane versioning in VersionID. This
is the case so far and once we see expection we can always add code
for this.

This also adds peppermint os.