hookstate,ifacestate: support snapctl set/get slot and plug attrs (step 3) (#2251)
snapctl - extended get and set subcommands to allow getting/setting of slot/plug attributes from the hooks. Initialize attributes with slot and plug attributes from the yaml when the hooks are executed.

The feature is still WIP. Attributes are not yet really used when connecting the interface and are not persisted. These will be implemented in separate PRs.